[DEV-1047] fixed the Dynamic class loader
