+++ core/src/main/scala/com/raphtory/internals/serialisers/DynamicClassLoader.scala
@@ -16,7 +16,7 @@ private[serialisers] class DynamicClassLoader(parent: ClassLoader, storage: Conc
 
   @throws[ClassNotFoundException]
   override def findClass(name: String): Class[_] =
-    Try(super.loadClass(name)) match {
+    Try(parent.loadClass(name)) match {
       case Success(cls)                       => cls
       case Failure(t: ClassNotFoundException) =>
         val clzBytes = storage.get(name)
